Change unit price into integer

+++ database/init_db.py
@@ -133,7 +133,7 @@ def init_database():
             task_id INTEGER NOT NULL,
             supplier_id INTEGER NOT NULL,
             pr_item_id INTEGER NOT NULL,
-            unit_price REAL,
+            unit_price INTEGER,
             stock_availability TEXT,
             cert TEXT,
             lead_time TEXT,
